requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
protobuf==3.12.2
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
